fixed minor bug with different operand types in the order model in the checkout app

+++ checkout/models.py
@@ -68,7 +68,7 @@ class Order(models.Model):
 
         if self.discount:
             discount_percentage = self.discount.discount
-            discount_value = self.order_total * Decimal(discount_percentage / 100)
+            discount_value = self.order_total * (discount_percentage / 100)
         else:
             discount_value = 0
